30_10 v5

+++ astec/database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('first_name');
             $table->string('last_name');
             $table->string('email')->unique();
-            $table->integer('cpf')->unique();
-            $table->integer('phone')->unique();
+            $table->bigInteger('cpf')->unique();
+            $table->bigInteger('phone')->unique();
             //$table->string('password');
             //$table->rememberToken();
             $table->timestamps();
